# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Open Agent Harness API
  version: 0.1.0
  summary: Headless distributed agent runtime API
  description: >
    Open Agent Harness is a headless agent runtime for multi-workspace,
    distributed usage. Identity, membership, and authentication are expected
    to be managed by an external service or upstream gateway. This document
    defines the HTTP API for workspace management, session messaging, run
    lifecycle, action invocation, and SSE event streaming.
servers:
  - url: http://localhost:8787/api/v1
    description: Local development
...
paths:
...
  /runtimes:
    get:
      tags: [Workspaces]
      summary: List workspace runtimes
      operationId: listWorkspaceRuntimes
      responses:
        '200':
          description: Workspace runtime list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceRuntimeList'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /runtimes/upload:
    post:
      tags: [Workspaces]
      summary: Upload workspace runtime
      operationId: uploadWorkspaceRuntime
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - in: query
          name: overwrite
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Workspace runtime uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceRuntimeUploadResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /runtimes/{runtimeName}:
    delete:
      tags: [Workspaces]
      summary: Delete workspace runtime
      operationId: deleteWorkspaceRuntime
      parameters:
        - in: path
          name: runtimeName
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '204':
          description: Workspace runtime deleted
        default:
          $ref: '#/components/responses/ErrorResponse'
  /model-providers:
    get:
      tags: [Models]
      summary: List supported model providers
      operationId: listModelProviders
      responses:
        '200':
          description: Supported model provider list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelProviderList'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /workspaces/{workspaceId}:
    get:
      tags: [Workspaces]
      summary: Get workspace
      operationId: getWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Workspace metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Workspaces]
      summary: Delete workspace
      operationId: deleteWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '204':
          description: Workspace deleted
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Timestamp:
      type: string
      format: date-time
    JsonObject:
      type: object
      additionalProperties: true
...
    WorkspaceRuntime:
      type: object
      required: [name]
      properties:
        name:
          type: string
    WorkspaceRuntimeList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceRuntime'
    WorkspaceRuntimeUploadResponse:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ModelProvider:
      type: object
      required:
        - id
        - packageName
        - description
        - requiresUrl
        - useCases
      properties:
        id:
          type: string
          enum: [openai, openai-compatible]
        packageName:
          type: string
        description:
          type: string
        requiresUrl:
          type: boolean
        useCases:
          type: array
          items:
            type: string
      additionalProperties: false
    ModelProviderList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ModelProvider'
      additionalProperties: false
    Workspace:
      type: object
      required:
        - id
        - name
        - rootPath
        - executionPolicy
        - status
        - kind
        - readOnly
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        externalRef:
          type: string
        name:
          type: string
        runtime:
          type: string
          description: Preferred field for the workspace runtime name.
        rootPath:
          type: string
        executionPolicy:
          type: string
          enum: [local, container, remote_runner]
        status:
          type: string
          enum: [active, archived, disabled]
        kind:
          type: string
          enum: [project]
        readOnly:
          type: boolean
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          $ref: '#/components/schemas/JsonObject'
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'